backend/requirements.txt
# Core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0

# Database
pymongo==4.6.0

# Authentication & Security
bcrypt==4.1.2
passlib==1.7.4
PyJWT==2.8.0

# Async tasks
celery==5.3.4
flower==2.0.1

# API & Data handling
marshmallow==3.20.2
requests==2.31.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3

# Development
pytest==7.4.4
pytest-flask==1.3.0
flake8==7.0.0
black==23.12.1